Fixing Music player container
