+++ artist.ts
